requirements.txt
boto3==1.35.21
botocore==1.35.21
firebase-admin==6.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.2
numpy==2.1.3
pydantic==2.9.2
requests==2.32.3
urllib3==2.2.3
